Better managment of player state
